# sisauth/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Application API document",
        "version": "1.0.0"
    },
    "paths": {
        "/api/addresses/{id}/details": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get details of an address",
                "description": "Returns details of a specific address",
                "operationId": "showAddressDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addresses/{uid}/user-details": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get addresses of a user",
                "description": "Returns addresses of a specific user",
                "operationId": "showUserAddresses",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Address"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addresses/{uid}/new": {
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Create a new address",
                "description": "Creates a new address associated with the specified user",
                "operationId": "storeAddress",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addresses/{id}/update": {
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update an existing address",
                "description": "Updates an existing address with the provided details",
                "operationId": "updateAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data to update the address",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "logradouro": {
                                        "description": "Street name",
                                        "type": "string"
                                    },
                                    "numero": {
                                        "description": "House number",
                                        "type": "string"
                                    },
                                    "bairro": {
                                        "description": "Neighborhood",
                                        "type": "string"
                                    },
                                    "complemento": {
                                        "description": "Address complement",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "cep": {
                                        "description": "Postal code",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "Associated user ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found"
                    },
                    "422": {
                        "description": "Invalid data provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addresses/{id}/delete": {
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete an existing address",
                "description": "Deletes an existing address by ID",
                "operationId": "deleteAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Address deleted successfully"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "title": "Address",
                "description": "Address model",
                "properties": {
                    "id": {
                        "description": "Address ID",
                        "type": "integer",
                        "example": 1
                    },
                    "logradouro": {
                        "description": "Street name",
                        "type": "string",
                        "example": "123 Main Street"
                    },
                    "numero": {
                        "description": "House number",
                        "type": "string",
                        "example": "101"
                    },
                    "bairro": {
                        "description": "Neighborhood",
                        "type": "string",
                        "example": "Downtown"
                    },
                    "complemento": {
                        "description": "Address complement",
                        "type": "string",
                        "example": "Apartment 5A"
                    },
                    "cep": {
                        "description": "Postal code",
                        "type": "string",
                        "example": "12345-678"
                    },
                    "user_id": {
                        "description": "Associated user ID",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}